pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "nostrpy"
authors = [
  { name="Richard Li", email="lifuzu+github@gmail.com" },
]
description = "Python CLI for Nostr"
urls = { Homepage = "https://github.com/nostrpy/nostrcli" }
readme = "README.md"
requires-python = ">=3.7.0"
dependencies = [
    "cffi>=1.15.0",
    "cryptography>=38.0.4",
    "pycparser>=2.21",
    "python-socks>=2.1.1",
    "secp256k1>=0.14.0",
    "websocket-client>=1.4.2",
    "click>=8.1.3",
]
license = {file = "LICENSE"}
classifiers=[
    'Operating System :: POSIX :: Linux',
    'Operating System :: Microsoft :: Windows',
    'Operating System :: MacOS :: MacOS X',
]
dynamic=["version"]

[tool.setuptools_scm]
write_to = "nostr/_version.py"

[project.optional-dependencies]
test = [
  "pytest >=7.2.0",
  "pytest-cov[all]"
]

[project.scripts]
nostr = "nostr.cli:cli"

[tool.black]
line-length = 88
